Got rid of annoying warnings during win64 build
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@7071 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/com/array.h
@@ -507,7 +507,7 @@ public:
 
 #else
 
-        SAFEARRAYBOUND bound = { aNewSize, 0 };
+        SAFEARRAYBOUND bound = { (ULONG)aNewSize, 0 };
         m.arr = SafeArrayCreate (VarType(), 1, &bound);
         AssertReturn (m.arr != NULL, false);
 
@@ -1000,7 +1000,7 @@ public:
 
 #else
 
-        SAFEARRAYBOUND bound = { aNewSize, 0 };
+        SAFEARRAYBOUND bound = { (ULONG)aNewSize, 0 };
         m.arr = SafeArrayCreateEx (VT_UNKNOWN, 1, &bound,
                                    (PVOID) &_ATL_IIDOF (I));
         AssertReturn (m.arr != NULL, false);
